+++ src/components/Print/index.ts
@@ -1,6 +1,12 @@
-export { PrintBin } from './PrintBin';
-export { PrintLayout } from './PrintLayout';
-export { PrintListEmpty } from './PrintListEmpty';
-export { PrintListSummary } from './PrintListSummary';
-export { SortOrderConfig } from './SortOrderConfig';
+// Re-export from new location for backward compatibility
+// TODO: Remove in Phase 6 - update all imports to use features/print-export
+export {
+  PrintBin,
+  PrintLayout,
+  PrintListEmpty,
+  PrintListSummary,
+  SortOrderConfig,
+} from '../../features/print-export';
+
+// SplitPreview remains local (not part of print-export feature)
 export { SplitPreview } from './SplitPreview';
