+++ src/features/print-export/components/PrintListSummary.tsx
@@ -1,4 +1,4 @@
-import { formatPrintTime, formatCost } from '../../utils/printEstimates';
+import { formatPrintTime, formatCost } from '../utils/printEstimates';
 
 interface PrintListSummaryProps {
   totalBins: number;
