+++ src/features/print-export/components/SortOrderConfig.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
-import { Checkbox } from '../../shared/components/Checkbox';
-import type { BinListSortOrder, SortFieldConfig, BinSortField } from '../../core/store/settings';
-import { SORT_FIELD_LABELS } from '../../core/store/settings';
+import { Checkbox } from '../../../shared/components/Checkbox';
+import type { BinListSortOrder, SortFieldConfig, BinSortField } from '../../../core/store/settings';
+import { SORT_FIELD_LABELS } from '../../../core/store/settings';
 
 interface SortOrderConfigProps {
   sortOrder: BinListSortOrder;
