+++ src/test/components/PrintBin.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
-import { PrintBin } from '../../components/Print/PrintBin';
+import { PrintBin } from '../../components/Print';
 import type { Bin, Category, Drawer } from '../../core/types';
 import type { PrintViewSettings } from '../../core/store/settings';
